refactor: move fetching staff schedules to shared MediatR query
