+++ Backend/Appointments/Appointments.API/Features/GetStaffAvailability/GetStaffAvailabilityHandler.cs
@@ -1,65 +1,31 @@
-using Appointments.API.Extensions;
-using Appointments.Domain.Common;
-using Appointments.Domain.Constants;
+using Appointments.API.Features.Shared.GetStaffSchedule;
 using Appointments.Domain.Enums;
-using Appointments.Domain.Interfaces;
 using Appointments.Infrastructure.Connection;
 using Dapper;
-using Google.Protobuf;
-using Grpc.Core;
-using InnoClinic.Contracts.Grpc;
 using InnoClinic.Core.Common;
 using MediatR;
 
 namespace Appointments.API.Features.GetStaffAvailability;
 
 public class GetStaffAvailabilityHandler(
-    ICacheService cacheService,
-    StaffScheduleSyncService.StaffScheduleSyncServiceClient profilesClient,
-    ISqlConnectionFactory connectionFactory,
-    ILogger<GetStaffAvailabilityHandler> logger)
+    ISender sender,
+    ISqlConnectionFactory connectionFactory)
     : IRequestHandler<GetStaffAvailabilityQuery, Result<StaffAvailabilityResponse>>
 {
     public async Task<Result<StaffAvailabilityResponse>> Handle(
         GetStaffAvailabilityQuery request, 
         CancellationToken cancellationToken)
     {
-        var redisKey = CacheConstants.MedicalStaffScheduleKey(request.MedicalStaffId);
+        var scheduleResult = await sender.Send(
+            new GetStaffScheduleQuery(request.MedicalStaffId), 
+            cancellationToken);
 
-        var scheduleJson = await cacheService.GetOrSetStringAsync(
-            redisKey,
-            async (ct) =>
-            {
-                logger.LogCacheMissFallback(request.MedicalStaffId);
-
-                try
-                {
-                    var fallbackSchedule = await profilesClient.GetStaffProfileAsync(
-                        new GetStaffProfileRequest { MedicalStaffId = request.MedicalStaffId.ToString() },
-                        cancellationToken: ct);
-
-                    var json = JsonFormatter.Default.Format(fallbackSchedule);
-                    logger.LogCacheRepopulated(request.MedicalStaffId);
-                    return json;
-                }
-                catch (RpcException ex) when (ex.StatusCode == StatusCode.NotFound)
-                {
-                    return null;
-                }
-                catch (Exception ex)
-                {
-                    logger.LogFallbackError(ex, request.MedicalStaffId);
-                    throw;
-                }
-            },
-            cancellationToken: cancellationToken);
-
-        if (string.IsNullOrEmpty(scheduleJson))
+        if (scheduleResult.IsFailure)
         {
-            return AppointmentErrors.MedicalStaffNotFound(request.MedicalStaffId);
+            return scheduleResult.Error;
         }
 
-        var schedule = JsonParser.Default.Parse<SyncStaffProfileRequest>(scheduleJson);
+        var schedule = scheduleResult.Value;
 
         using var connection = connectionFactory.CreateConnection();
