+++ Backend/Appointments/Appointments.API/Features/Shared/GetStaffSchedule/GetStaffScheduleQuery.cs
@@ -1,0 +1,7 @@
+using InnoClinic.Contracts.Grpc;
+using InnoClinic.Core.Common;
+using MediatR;
+
+namespace Appointments.API.Features.Shared.GetStaffSchedule;
+
+public record GetStaffScheduleQuery(Guid MedicalStaffId) : IRequest<Result<SyncStaffProfileRequest>>;
